fix: 유효성체크 훅, 유틸 함수로 변경
